python/requirements.txt
openai~=1.10.0
python-dotenv~=1.0.1